+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this tool will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.4]
+### Changed
+- Small fixes
+- Package name is now displayed
+- Success message changed from 'OK' to 'Merged'
+
 ## [1.1.3]
 ### Added
 - The dependency type is now displayed